mac80211: Allow preamble configuration
Add a new wireless config parameter short_preamble=0|1 to enable
usage of short preambles. Default is to only allow long preambles
as before.

Even if short_preamble is set to 1 hostapd will take care that
short preambles are disabled as soon as a STA associates that
cannot handle short preambles.

Signed-off-by: Helmut Schaa <helmut.schaa@googlemail.com>