AA: brcm47xx: backport r34150: remove ssb-gige package
CONFIG_SSB_DRIVER_GIGE is set in the target config of the brcm47xx target, so it will always be activated and the kmod-ssb-gige is not needed any more.
The kmod-ssb-gige package is not generated in an build as it does not contain any files and the makes the image builder fail when generating a image integrating this package.